+++ src/storage/store/store_sync/filesystem_store.rs
@@ -6,6 +6,7 @@ use crate::{
     array::MaybeBytes,
     byte_range::{ByteOffset, ByteRange},
     storage::{
+        store_lock::{DefaultStoreLocks, StoreKeyMutex, StoreLocks},
         store_set_partial_values, ListableStorageTraits, ReadableStorageTraits,
         ReadableWritableStorageTraits, StorageError, StoreKey, StoreKeyError, StoreKeyRange,
         StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix, StorePrefixes,
@@ -55,17 +56,30 @@ pub struct FilesystemStore {
     sort: bool,
     readonly: bool,
     files: Mutex<HashMap<StoreKey, Arc<RwLock<()>>>>,
+    locks: StoreLocks,
 }
 
 impl FilesystemStore {
     /// Create a new file system store at a given `base_path`.
     ///
     /// # Errors
-    ///
     /// Returns a [`FilesystemStoreCreateError`] if `base_directory`:
     ///   - is not valid, or
     ///   - it points to an existing file rather than a directory.
     pub fn new<P: AsRef<Path>>(base_path: P) -> Result<Self, FilesystemStoreCreateError> {
+        Self::new_with_locks(base_path, Arc::new(DefaultStoreLocks::default()))
+    }
+
+    /// Create a new file system store at a given `base_path` with non-default store locks.
+    ///
+    /// # Errors
+    /// Returns a [`FilesystemStoreCreateError`] if `base_directory`:
+    ///   - is not valid, or
+    ///   - it points to an existing file rather than a directory.
+    pub fn new_with_locks<P: AsRef<Path>>(
+        base_path: P,
+        store_locks: StoreLocks,
+    ) -> Result<Self, FilesystemStoreCreateError> {
         let base_path = base_path.as_ref().to_path_buf();
         if base_path.to_str().is_none() {
             return Err(FilesystemStoreCreateError::InvalidBasePath(base_path));
@@ -87,6 +101,7 @@ impl FilesystemStore {
             sort: false,
             readonly,
             files: Mutex::default(),
+            locks: store_locks,
         })
     }
 
@@ -316,7 +331,11 @@ impl WritableStorageTraits for FilesystemStore {
     }
 }
 
-impl ReadableWritableStorageTraits for FilesystemStore {}
+impl ReadableWritableStorageTraits for FilesystemStore {
+    fn mutex(&self, key: &StoreKey) -> Result<StoreKeyMutex, StorageError> {
+        Ok(self.locks.mutex(key))
+    }
+}
 
 impl ListableStorageTraits for FilesystemStore {
     fn list(&self) -> Result<StoreKeys, StorageError> {